+++ apps/api/src/notifications/mod.rs
@@ -9,7 +9,7 @@ use rocket::{State, get, post, tokio::sync::Mutex};
 #[get("/notifications/<device_id>?<from_timestamp>")]
 pub async fn get_notifications(
     device_id: DeviceIdParam,
-    from_timestamp: Option<u32>,
+    from_timestamp: Option<u64>,
     client: &State<Mutex<NotificationsClient>>,
 ) -> Result<ApiResponse<Vec<InAppNotification>>, ApiError> {
     Ok(client.lock().await.get_notifications(&device_id.0, from_timestamp)?.into())
